Make EditPlaylist dialog content reusable
